Extract domain objects into their own file
